Unsupervised Clustering: TSNE, DBSCAN, Hierarchical Clustering
For a dataset loaded as .pkl file, first a dendogram is plotted (hierarchical clustering). When dendogram is cut at a height, for each cluster produced, 2 plots are created.

First is a TSNE plot based on the raw permission group each APK contains. Second is a TSNE plot based on the label assigned via DBSCAN algorithm.